requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.4
celery==5.6.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
cryptography==45.0.4
defusedxml==0.7.1
Django==4.2.21
django-celery-beat==2.8.1
django-timezone-field==7.2.1
django_celery_results==2.6.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
hijridate==2.5.0
idna==3.10
isodate==0.7.2
kombu==5.6.1
lxml==5.4.0
openpyxl==3.1.5
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.52
pycparser==2.22
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==7.0.1
reportlab==4.4.7
requests==2.32.4
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
xlsxwriter==3.2.5
zeep==4.3.1
